+++ ane/storiesane/eval_cached_other.go
@@ -8,11 +8,11 @@ import "fmt"
 func (e *Engine) ResetCache() {}
 
 // EvalNextToken is unavailable on non-darwin platforms.
-func (e *Engine) EvalNextToken(token uint16) ([]float32, error) {
+func (e *Engine) EvalNextToken(token int32) ([]float32, error) {
 	return nil, fmt.Errorf("storiesane eval next token: unavailable on this platform")
 }
 
 // EvalPrefill is unavailable on non-darwin platforms.
-func (e *Engine) EvalPrefill(tokens []uint16) ([]float32, error) {
+func (e *Engine) EvalPrefill(tokens []int32) ([]float32, error) {
 	return nil, fmt.Errorf("storiesane eval prefill: unavailable on this platform")
 }
